[svn-r11886] Purpose:     Code cleanup
Description:
    Check in some of the code cleanups from working on the external link
support.  (This doesn't include any of the external link features)

Platforms tested:
    FreeBSD 4.11 (sleipnir)
    Mac OSX.4 (amazon)
    Linux 2.4

+++ hl/src/H5LT.c
@@ -2110,8 +2110,8 @@ print_enum(hid_t type, char* str, int indt)
     dst_size = H5Tget_size(native);
     
     /* Get the names and raw values of all members */
-    name = (char**)calloc(nmembs, sizeof(char *));
-    value = (unsigned char*)calloc(nmembs, MAX(dst_size, super_size));
+    name = (char**)calloc((size_t)nmembs, sizeof(char *));
+    value = (unsigned char*)calloc((size_t)nmembs, MAX(dst_size, super_size));
 
     for (i = 0; i < nmembs; i++) {
 	if((name[i] = H5Tget_member_name(type, i))==NULL)
